Make IPMIService for IPMI pool calls
